chumby8: enable DMA
This is step 1 in preparation for getting sound working. DMA has now
been tested with the touchscreen and it works.

Currently a bunch of IRQ errors relating to the DMA controller will be
printed at startup. These errors are harmless. I'll try to figure out
how to suppress them soon.