chumby8: add headphone detection and microphone support
The headphone detection was simple; we just needed the GPIO to be hooked
up. This also required filling out the widgets, because it specifically
looks for a widget called "Headphones".

Also hook up MICBIAS in the routing in order for the microphone to work
properly. This commit also adds the rest of the routing similar to other
usages of simple-audio-card.