chumby8: enable LCD controller and GPU in device tree
Note that the armada-drm driver works differently from the others, in
that it doesn't use the normal CMA mechanism. So reserve memory for the
armada driver in the device tree.

Remove the reserved memory we were using for the simpledrm framebuffer.
Note that u-boot doesn't let us customize the framebuffer location in
RAM, so it has to be manually copied to a location in the defined range
in order for linux to not clobber the display while loading. This is
easy to do inside of u-boot.