libdrm-armada: new package
libdrm-armada is an Etnaviv libdrm buffer object management module.
It is used with xf86-driver-video-armada, which is Xorg driver for Etnaviv.

This patch was originally written by Evgeniy Didin and shared to the
buildroot mailing list. However, it never got merged to mainline
buildroot. Doug Brown updated the patch for the needs of the Chumby 8.

Original patch:
https://lists.buildroot.org/pipermail/buildroot/2018-February/515431.html

Signed-off-by: Doug Brown <doug@schmorgal.com>