chumby8: start X with white screen
This changes the default screen to white instead of black, which makes
the display more "Chumby-esque" at boot before a program has loaded.

There is still a brief moment where the screen goes black as soon as
Xorg starts. I'm not sure what's causing this, but unfortunately that
means we don't have a completely smooth transition between the splash
screen and the white Xorg background.