chumby8: make simple automatic time save daemon
This is done entirely as a shell script. The daemon runs very early in
the boot process. It loads the time from our fake "RTC" in the
cryptoprocessor combined with the file we save in /etc, if the file
exists. If the file doesn't exist, it waits until the date/time gets set
automatically by NTP.

Once a good date/time has been received from NTP, it saves the current
time and the cryptoprocessor's uptime to a file so it can be retrieved
at the next startup.

Note that the automatic save does result in a quick remount of the SD
card. I'm considering moving some of these settings files to a different
settings partition that will be kept read/write at all times.