chumby8: switch to new u-boot with default quirk for card reader
Now that we have a userspace daemon handling the card reader, we can
change the default boot command to disable the card reader by default.
This prevents a dummy sda device from showing up if you boot without a
card inserted. It does mean the card won't be available until the
userspace utility enables it after detecting the card, which is a very
small tradeoff.

Also, switch to downloading the tarball from GitHub instead of cloning
the entire repository.